Photographer/Photographer/backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
psycopg2-binary>=2.9.1,<3.0.0
sqlalchemy>=1.4.23,<2.0.0
firebase-admin>=5.0.0,<6.0.0
python-socketio>=5.4.0,<6.0.0
google-auth>=2.3.0,<3.0.0
google-auth-oauthlib>=0.4.6,<0.5.0
requests>=2.26.0,<3.0.0
aiohttp>=3.8.0,<4.0.0